backend/tasks/src/requirements.txt
alembic==1.8.1
amqp==5.1.1
aniso8601==9.0.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0.post0
grpcio==1.50.0
grpcio-tools==1.50.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
prompt-toolkit==3.0.32
protobuf==4.21.9
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.43
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
